+++ aws-runtime/aws-config/common/test/aws/sdk/kotlin/runtime/config/AwsSdkSettingTest.kt
@@ -1,5 +1,6 @@
 package aws.sdk.kotlin.runtime.config
 
+import aws.sdk.kotlin.runtime.config.retries.RetryMode
 import aws.smithy.kotlin.runtime.util.PlatformEnvironProvider
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -49,6 +50,24 @@ class AwsSdkSettingTest {
 
         val actual = AwsSdkSetting.AwsEc2MetadataDisabled.resolve(testPlatform)
         assertEquals(true, actual)
+    }
+
+    @Test
+    fun itResolvesMaxAttemptsFromEnvironmentVariables() {
+        val expected = 5
+        val testPlatform = mockPlatform(mapOf("AWS_MAX_ATTEMPTS" to expected.toString()), mapOf())
+
+        val actual = AwsSdkSetting.AwsMaxAttempts.resolve(testPlatform)
+        assertEquals(expected, actual)
+    }
+
+    @Test
+    fun itResolvesRetryModeFromSystemProperties() {
+        val expected = RetryMode.LEGACY
+        val testPlatform = mockPlatform(mapOf(), mapOf("aws.retryMode" to "legacy"))
+
+        val actual = AwsSdkSetting.AwsRetryMode.resolve(testPlatform)
+        assertEquals(expected, actual)
     }
 
     private fun mockPlatform(env: Map<String, String>, jvmProps: Map<String, String>): PlatformEnvironProvider {